requirements.txt
bleach==1.5.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
cycler==0.10.0
enum34==1.1.6
future==0.16.0
html5lib==0.9999999
idna==2.6
Keras==2.1.2
Markdown==2.6.10
matplotlib==2.1.1
numpy==1.13.3
oauthlib==2.0.6
pandas==0.21.1
protobuf==3.5.1
PyJWT==1.5.3
pyparsing==2.2.0
PySocks==1.6.8
python-dateutil==2.6.1
python-dotenv==0.7.1
python-twitter==3.3
pytz==2017.3
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
scipy==1.0.0
six==1.11.0
tensorflow==1.4.1
tensorflow-tensorboard==0.4.0rc3
twilio==6.10.0
urllib3==1.22
Werkzeug==0.13
